OVStyle adds supported escape sequences
Add escape sequences that tcell does not support to OVStyle.
Separate ovstyle.go from oviewer.go.
