+++ oviewer/oviewer.go
@@ -284,46 +284,6 @@ type Config struct {
 	Debug bool
 }
 
-// OVStyle represents a style in addition to the original style.
-type OVStyle struct {
-	// Background is a color name string.
-	Background string
-	// Foreground is a color name string.
-	Foreground string
-	// If true, add blink.
-	Blink bool
-	// If true, add bold.
-	Bold bool
-	// If true, add dim.
-	Dim bool
-	// If true, add italic.
-	Italic bool
-	// If true, add reverse.
-	Reverse bool
-	// If true, add underline.
-	Underline bool
-	// If true, add strike through.
-	StrikeThrough bool
-	// If true, add overline (not yet supported).
-	OverLine bool
-	// If true, sub blink.
-	UnBlink bool
-	// If true, sub bold.
-	UnBold bool
-	// If true, sub dim.
-	UnDim bool
-	// If true, sub italic.
-	UnItalic bool
-	// If true, sub reverse.
-	UnReverse bool
-	// If true, sub underline.
-	UnUnderline bool
-	// If true, sub strike through.
-	UnStrikeThrough bool
-	// if true, sub underline (not yet supported).
-	UnOverLine bool
-}
-
 var (
 	// MemoryLimit is a number that limits the chunks to load into memory.
 	MemoryLimit int
@@ -841,84 +801,6 @@ func (root *Root) debugMessage(msg string) {
 		return
 	}
 	log.Printf("%s:%s", root.Doc.FileName, msg)
-}
-
-// ToTcellStyle convert from ovStyle to tcell style.
-func ToTcellStyle(s OVStyle) tcell.Style {
-	style := tcell.StyleDefault
-	style = style.Background(tcell.GetColor(s.Background))
-	style = style.Foreground(tcell.GetColor(s.Foreground))
-	style = style.Blink(s.Blink)
-	style = style.Bold(s.Bold)
-	style = style.Dim(s.Dim)
-	style = style.Italic(s.Italic)
-	style = style.Reverse(s.Reverse)
-	style = style.Underline(s.Underline)
-	style = style.StrikeThrough(s.StrikeThrough)
-
-	return style
-}
-
-func applyStyle(style tcell.Style, s OVStyle) tcell.Style {
-	if s.Background != "" {
-		style = style.Background(tcell.GetColor(s.Background))
-	}
-	if s.Foreground != "" {
-		style = style.Foreground(tcell.GetColor(s.Foreground))
-	}
-	if s.Blink {
-		style = style.Blink(true)
-	}
-	if s.Bold {
-		style = style.Bold(true)
-	}
-	if s.Dim {
-		style = style.Dim(true)
-	}
-	if s.Italic {
-		style = style.Italic(true)
-	}
-	if s.Reverse {
-		style = style.Reverse(true)
-	}
-	if s.Underline {
-		style = style.Underline(true)
-	}
-	if s.StrikeThrough {
-		style = style.StrikeThrough(true)
-	}
-	if s.UnBlink {
-		style = style.Blink(false)
-	}
-	if s.UnBold {
-		style = style.Bold(false)
-	}
-	if s.UnDim {
-		style = style.Dim(false)
-	}
-	if s.UnItalic {
-		style = style.Italic(false)
-	}
-	if s.UnReverse {
-		style = style.Reverse(false)
-	}
-	if s.UnUnderline {
-		style = style.Underline(false)
-	}
-	if s.UnStrikeThrough {
-		style = style.StrikeThrough(false)
-	}
-	if s.OverLine {
-		// tcell does not support overline.
-		// style = style.OverLine(true)
-		style = style.Underline(true)
-	}
-	if s.UnOverLine {
-		// tcell does not support unOverline.
-		// style = style.UnOverLine(false)
-		style = style.Underline(false)
-	}
-	return style
 }
 
 // mergeGeneral overwrites a general structure with a struct.
